# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: K&S Solar Energy API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /complaints:
    get:
      operationId: getComplaints
      tags: [complaints]
      summary: Get complaints (own or all for admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of complaints
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Complaint"
    post:
      operationId: createComplaint
      tags: [complaints]
      summary: Submit a complaint
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateComplaintRequest"
      responses:
        "201":
          description: Complaint submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
        "401":
          description: Unauthorized
  /complaints/guest:
    post:
      operationId: createGuestComplaint
      tags: [complaints]
      summary: Submit a complaint without an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateComplaintRequest"
      responses:
        "201":
          description: Complaint submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
        "400":
          description: Bad request
  /complaints/{id}:
    get:
      operationId: getComplaintById
      tags: [complaints]
      summary: Get complaint by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Complaint details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
        "403":
          description: Forbidden
        "404":
          description: Not found
    patch:
      operationId: updateComplaint
      tags: [complaints]
      summary: Update complaint status and/or technician info (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateComplaintRequest"
      responses:
        "200":
          description: Complaint updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
        "403":
          description: Forbidden
        "404":
          description: Not found
  /complaints/{id}/status:
    patch:
      operationId: updateComplaintStatus
      tags: [complaints]
      summary: Update complaint status (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusRequest"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
        "403":
          description: Forbidden
...
  /complaints/{id}/technicians:
    put:
      operationId: assignComplaintTechnicians
      tags: [complaints]
      summary: Assign multiple technicians to a complaint (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignTechniciansRequest"
      responses:
        "200":
          description: Technicians assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Complaint"
        "403":
          description: Forbidden
        "404":
          description: Not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UpdateStatusRequest:
      type: object
      properties:
        status:
          type: string
      required:
        - status
...
    CreateComplaintRequest:
      type: object
      properties:
        subject:
          type: string
          enum: [Hybrid System, OnGrid System, Off-Grid System, Tubewell System]
        customerName:
          type: string
        phone:
          type: string
        address:
          type: string
        message:
          type: string
        bookingId:
          type: string
      required:
        - subject
        - customerName
        - phone
        - address
        - message
...
    AssignTechniciansRequest:
      type: object
      properties:
        technicianIds:
          type: array
          items:
            type: string
      required:
        - technicianIds
...
    UpdateComplaintRequest:
      type: object
      properties:
        status:
          type: string
          enum: [submitted, in_progress, resolved, closed]
        technicianName:
          type: string
          nullable: true
        technicianPhone:
          type: string
          nullable: true
        technicianId:
          type: string
          nullable: true
    Complaint:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        bookingId:
          type: string
        subject:
          type: string
        customerName:
          type: string
        phone:
          type: string
        address:
          type: string
        message:
          type: string
        status:
          type: string
        technicianName:
          type: string
          nullable: true
        technicianPhone:
          type: string
          nullable: true
        technicianId:
          type: string
          nullable: true
        technicianIds:
          type: array
          items:
            type: string
        statusHistory:
          type: array
          nullable: true
          items:
            type: object
            properties:
              status:
                type: string
              changedAt:
                type: string
            required:
              - status
              - changedAt
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - subject
        - customerName
        - phone
        - address
        - message
        - status
        - technicianIds
        - createdAt
...