# lib/api-zod/src/generated/types/complaint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * K&S Solar Energy API
 * OpenAPI spec version: 0.1.0
 */
import type { ComplaintStatusHistoryItem } from './complaintStatusHistoryItem';

export interface Complaint {
  id: string;
  userId: string;
  bookingId?: string;
  subject: string;
  customerName: string;
  phone: string;
  address: string;
  message: string;
  status: string;
  technicianName?: string | null;
  technicianPhone?: string | null;
  technicianId?: string | null;
  technicianIds: string[];
  statusHistory?: ComplaintStatusHistoryItem[] | null;
  createdAt: Date;
}